Move format() into a separate file
